gurobi/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
anyqt=0.0.10=py_1
astroid=2.3.3=py37_0
attrs=19.3.0=py_0
backcall=0.1.0=py37_0
blas=1.0=openblas
bleach=3.1.0=py37_0
bottleneck=1.3.2=py37hc1659b7_0
ca-certificates=2020.1.1=0
cachecontrol=0.12.6=py_0
certifi=2019.11.28=py37_0
cffi=1.14.0=py37hd463f26_0
chardet=3.0.4=py37hc8dfbb8_1006
commonmark=0.9.1=py_0
cryptography=2.8=py37hb09aad4_2
cycler=0.10.0=py37_0
dbus=1.13.12=h746ee38_0
decorator=4.4.1=py_0
defusedxml=0.6.0=py_0
docutils=0.16=py37hc8dfbb8_1
entrypoints=0.3=py37_0
et_xmlfile=1.0.1=py_1001
expat=2.2.6=he6710b0_0
flake8=3.7.9=py37_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
future=0.18.2=py37hc8dfbb8_1
glib=2.63.1=h5a9c865_0
gmp=6.1.2=h6c8ec71_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
gurobi=9.0.0=py37_0
icu=58.2=h9c2bf20_1
idna=2.9=py_1
importlib_metadata=1.4.0=py37_0
ipdb=0.12.3=py_0
ipykernel=5.1.4=py37h39e3cac_0
ipython=7.11.1=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
isort=4.3.21=py37_0
jdcal=1.4.1=py_0
jedi=0.16.0=py37_0
jeepney=0.4.3=py_0
jinja2=2.10.3=py_0
joblib=0.14.1=py_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py37_0
jupyter=1.0.0=py37_7
jupyter_client=5.3.4=py37_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.1=py37_0
keyring=21.1.1=py37_0
keyrings.alt=3.1=py_0
kiwisolver=1.1.0=py37he6710b0_0
lazy-object-proxy=1.4.3=py37h7b6447c_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=11_openblas
libcblas=3.8.0=11_openblas
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libllvm8=8.0.1=hc9558a2_0
libopenblas=0.3.6=h5a2b251_2
libpng=1.6.37=hbc83047_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
llvmlite=0.31.0=py37h5202443_1
lockfile=0.12.2=py_1
markupsafe=1.1.1=py37h7b6447c_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37hef1b27d_0
mccabe=0.6.1=py37_1
mistune=0.8.4=py37h7b6447c_0
more-itertools=8.0.2=py_0
msgpack-python=1.0.0=py37hc9558a2_0
mypy=0.740=py_0
mypy_extensions=0.4.1=py37_0
nbconvert=5.6.1=py37_0
nbformat=5.0.4=py_0
ncurses=6.1=he6710b0_1
networkx=2.4=py_1
notebook=6.0.3=py37_0
numba=0.48.0=py37hb3f55d8_0
numpy=1.18.1=py37h94c655d_0
numpy-base=1.18.1=py37h2f8d375_0
openpyxl=3.0.3=py_0
openssl=1.1.1e=h7b6447c_0
opentsne=0.3.12=py37h516909a_0
orange-canvas-core=0.1.11=py_0
orange-widget-base=4.4.0=py_1
orange3=3.24.1=py37he1b5a44_0
orange3-network=1.5.0=py37hb3f55d8_0
packaging=20.3=py_0
pandas=1.0.1=py37h0573a6f_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.6.0=py_0
pcre=8.43=he6710b0_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pip=19.3.1=py37_0
pluggy=0.13.1=py37_0
prometheus_client=0.7.1=py_0
prompt_toolkit=3.0.2=py_0
psutil=5.6.7=py37h7b6447c_0
ptyprocess=0.6.0=py37_0
py=1.8.1=py_0
pycodestyle=2.5.0=py37_0
pycparser=2.20=py_0
pyflakes=2.1.1=py37_0
pygments=2.5.2=py_0
pylint=2.4.4=py37_0
pynndescent=0.3.3=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h05f1152_2
pyqtgraph=0.10.0=py37h28b3542_3
pyrsistent=0.15.7=py37h7b6447c_0
pysocks=1.7.1=py37hc8dfbb8_1
pytest=5.4.1=py37_0
python=3.7.6=h0371630_2
python-dateutil=2.8.1=py_0
python-louvain=0.13=py_0
python_abi=3.7=1_cp37m
pytz=2019.3=py_0
pyyaml=5.3=py37h7b6447c_0
pyzmq=18.1.0=py37he6710b0_0
qt=5.9.7=h5867ecd_1
qtconsole=4.6.0=py_1
readline=7.0=h7b6447c_5
requests=2.23.0=py37hc8dfbb8_1
scikit-learn=0.22.2.post1=py37hcdab131_0
scipy=1.3.2=py37he2b7bc3_0
secretstorage=3.1.2=py37hc8dfbb8_1
send2trash=1.5.0=py37_0
serverfiles=0.3.0=py_0
setuptools=44.0.0=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.13.0=py37_0
sqlite=3.30.1=h7b6447c_0
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
tk=8.6.8=hbc83047_0
tornado=6.0.3=py37h7b6447c_0
traitlets=4.3.3=py37_0
typed-ast=1.4.0=py37h7b6447c_0
typing_extensions=3.7.4.1=py37_0
urllib3=1.25.7=py37hc8dfbb8_1
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
wheel=0.33.6=py37_0
widgetsnbextension=3.5.1=py37_0
wrapt=1.11.2=py37h7b6447c_0
xlrd=1.2.0=py_0
xlsxwriter=1.2.8=py_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zeromq=4.3.1=he6710b0_3
zipp=0.6.0=py_0
zlib=1.2.11=h7b6447c_3
